chore(builder): cleans up the builder code a bit more
