+++ crates/builder/core/src/flashblocks/mod.rs
@@ -3,11 +3,20 @@
 mod best_txs;
 pub use best_txs::BestFlashblocksTxs;
 
+mod cell;
+pub use cell::{BlockCell, WaitForValue};
+
+mod resolve;
+pub use resolve::ResolvePayload;
+
+mod args;
+pub use args::BuildArguments;
+
+mod job;
+pub use job::BlockPayloadJob;
+
 mod generator;
-pub use generator::{
-    BlockCell, BlockPayloadJob, BlockPayloadJobGenerator, BuildArguments, ResolvePayload,
-    WaitForValue,
-};
+pub use generator::BlockPayloadJobGenerator;
 
 mod traits;
 pub use traits::PayloadBuilder;
@@ -21,8 +30,11 @@ pub use config::FlashblocksConfig;
 mod context;
 pub use context::{FlashblocksExtraCtx, OpPayloadBuilderCtx};
 
+mod info;
+pub use info::FlashblocksExecutionInfo;
+
 mod payload;
-pub use payload::FlashblocksExecutionInfo;
+pub use payload::OpPayloadBuilder;
 
 mod service;
 pub use service::FlashblocksServiceBuilder;
